requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
openai==0.27.6
deepl==1.14.0
pydub==0.25.1
python-dotenv==1.0.0
